organize docker integration (tests)
